+++ app/src/main/AndroidManifest.xml
@@ -12,7 +12,7 @@
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
         <activity
-            android:name=".Login"
+            android:name="Activities.Login"
             android:windowSoftInputMode="stateAlwaysVisible">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -25,7 +25,7 @@
             android:name="org.apache.http.legacy"
             android:required="false" />
 
-        <activity android:name=".SignUp"></activity>
+        <activity android:name="Activities.SignUp"></activity>
     </application>
 
 </manifest>
