+++ app/src/main/java/Activities/Login.java
@@ -1,16 +1,13 @@
-package com.cmput301f18t20.medicalphotorecord;
+package Activities;
 
 import android.content.Intent;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
-import android.util.Log;
 import android.view.View;
 import android.widget.Button;
 import android.widget.EditText;
 
-import com.searchly.jestdroid.DroidClientConfig;
-
-import java.util.ArrayList;
+import com.cmput301f18t20.medicalphotorecord.R;
 
 public class Login extends AppCompatActivity {
