+++ app/src/main/java/Controllers/ElasticsearchRecordsController.java
@@ -1,8 +1,9 @@
-package com.cmput301f18t20.medicalphotorecord;
+package Controllers;
 
 import android.os.AsyncTask;
 import android.util.Log;
 
+import com.cmput301f18t20.medicalphotorecord.Record;
 import com.searchly.jestdroid.DroidClientConfig;
 import com.searchly.jestdroid.JestClientFactory;
 import com.searchly.jestdroid.JestDroidClient;
@@ -14,15 +15,13 @@ import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
 
-import io.searchbox.client.JestClient;
 import io.searchbox.client.JestResult;
 import io.searchbox.core.DocumentResult;
 import io.searchbox.core.Index;
 import io.searchbox.core.Search;
-import io.searchbox.indices.CreateIndex;
 
 /* TODO CREDIT we will need to credit this to the lonelyTwitter lab guy */
-public class ElasticsearchProviderController {
+public class ElasticsearchRecordsController {
 
     static JestDroidClient client = null;
 
@@ -39,12 +38,12 @@ public class ElasticsearchProviderController {
         }
     }
 
-    public static class GetProviderTask extends AsyncTask<String, Void, ArrayList<Provider>>{
+    public static class GetRecordsTask extends AsyncTask<String, Void, ArrayList<Record>>{
         @Override
         //String instead of void to implement search
-        protected ArrayList<Provider> doInBackground(String... params) {
+        protected ArrayList<Record> doInBackground(String... params) {
             setClient();
-            ArrayList<Provider> Providers =new ArrayList<Provider>();
+            ArrayList<Record> Records=new ArrayList<Record>();
             String query = "{\n" +
                     //"    \"id\": \"myTemplateId\"," +
                     "    \"params\": {\n" +
@@ -54,32 +53,32 @@ public class ElasticsearchProviderController {
 
             Search search = new Search.Builder("")//query)
                     .addIndex("cmput301f18t20")
-                    .addType("Provider")
+                    .addType("Record")
                     .build();
             try {
                 JestResult result=client.execute(search);
 
                 if(result.isSucceeded()){
-                    List<Provider> ProviderList;
-                    ProviderList=result.getSourceAsObjectList(Provider.class);
-                    Providers.addAll(ProviderList);
+                    List<Record> RecordList;
+                    RecordList=result.getSourceAsObjectList(Record.class);
+                    Records.addAll(RecordList);
                 }
 
             }catch(IOException e){}
 
-            return Providers;
+            return Records;
         }
     }
 
-    public static class AddProviderTask extends AsyncTask<Provider, Void, Void>{
+    public static class AddRecordTask extends AsyncTask<Record, Void, Void>{
         @Override
-        protected Void doInBackground(Provider... params){
+        protected Void doInBackground(Record... params){
             setClient();
 
-            Provider Provider = params[0];
-            Index index=new Index.Builder(Provider)
+            Record Record = params[0];
+            Index index=new Index.Builder(Record)
                     .index("cmput301f18t20")
-                    .type("Provider")
+                    .type("Record")
                     .build();
 
             try {
@@ -92,6 +91,8 @@ public class ElasticsearchProviderController {
                 Log.d("Hello", "IOEXCEPTION");
             }
             return null;
+
         }
+
     }
 }
