+++ app/src/test/java/com/cmput301f18t20/medicalphotorecord/ProviderTests.java
@@ -5,6 +5,8 @@ import org.junit.Test;
 import java.util.ArrayList;
 import java.util.NoSuchElementException;
 
+import Exceptions.UserIDMustBeAtLeastEightCharactersException;
+
 import static junit.framework.TestCase.fail;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
